API uses nodemon and tests use ava
