+++ notary-api/package.json
@@ -7,10 +7,10 @@
   "scripts": {
     "build": "babel src -d dist",
     "serve": "node dist/index.js",
-    "start": "babel-node src/index.js",
-    "test": "scripts/test.sh",
-    "test:coverage": "nyc scripts/test.sh",
-    "test:coverage:ci": "nyc scripts/test.sh > coverage.lcov",
+    "start": "nodemon src/index.js",
+    "test": "ava --tap | faucet",
+    "test:watch": "ava --watch",
+    "test:coverage": "nyc npm test",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "publish": "scripts/publish.sh $npm_package_name"
@@ -46,6 +46,7 @@
     "winston": "^3.0.0"
   },
   "devDependencies": {
+    "ava": "^0.25.0",
     "babel-cli": "^6.26.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.7.0",
@@ -54,14 +55,27 @@
     "eslint": "^4.19.1",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.12.0",
+    "faucet": "0.0.1",
     "ganache-cli": "^6.1.3",
     "level-test": "^4.0.0",
-    "mocha": "^5.2.0",
+    "nodemon": "^1.18.9",
     "nyc": "12.0.2",
     "redis-mock": "^0.27.0",
     "sinon": "^6.1.3",
     "supertest": "^3.1.0",
     "testdouble": "^3.8.1"
+  },
+  "nodemonConfig": {
+    "exec": "clear;babel-node",
+    "watchedExtensions": [
+      "js"
+    ],
+    "ignore": [
+      "__mocks__/*",
+      "__test__/*",
+      "*.test.js"
+    ],
+    "delay": "7000"
   },
   "nyc": {
     "check-coverage": true,
@@ -73,6 +87,22 @@
     ],
     "exclude": [
       "src/utils/logger.js"
+    ]
+  },
+  "ava": {
+    "files": [
+      "test/**/*.test.js"
+    ],
+    "cache": true,
+    "concurrency": 5,
+    "failWithoutAssertions": false,
+    "tap": true,
+    "verbose": true,
+    "compileEnhancements": false,
+    "require": [
+      "babel-register",
+      "babel-polyfill",
+      "./test/setup"
     ]
   }
 }
